Update data-transform.qmd
Response to question: 
 "In a single pipeline for each condition, find all flights that meet the condition: Arrived more than two hours late, but didn’t leave late"

Code written in the suggested answers: 

`flights |> 
  filter(arr_delay >= 120 & dep_delay <= 0) |> view()`

This gives 0 results, however, I was able to manage to get code that gets results meeting the conditions:

`flights |> filter(arr_delay > 120) |> filter(dep_delay <= 0) |>view()`

By using the pipe to get to this answer, I first made it look thoughout the tibble for flights that had an arrival delay of more then two hours, then I filtered out of this tibble the flights that met the condition `dep_delay <= 0`

The reason I used `arr_delay >120` is because the question states more than 2 hours late, not at least 2 hours late. When using `>= 120`, you get the same result. 

This code results in :

[Issue.csv](https://github.com/user-attachments/files/16801488/Issue.csv)

Preview of the tibble:

![image](https://github.com/user-attachments/assets/57e3341f-1af0-4424-bf6d-942288ddc424)